Rename the package.
