+++ runtests.py
@@ -37,11 +37,11 @@ if not settings.configured:
         DATABASES=DATABASES,
         INSTALLED_APPS=(
             'django.contrib.contenttypes',
-            'sendgrid',
+            'django_sendgrid',
         ),
         SITE_ID=1,
         SECRET_KEY='this-is-just-for-tests-so-not-that-secret',
-        ROOT_URLCONF='sendgrid.urls',
+        ROOT_URLCONF='django_sendgrid.urls',
         TIME_ZONE='UTC',  # so we can switch USE_TZ on and off in-flight with postgres
         MIDDLEWARE_CLASSES=('django.middleware.csrf.CsrfViewMiddleware', )
     )
@@ -53,7 +53,7 @@ from django.test.utils import get_runner
 def run_tests():
     if hasattr(django, 'setup'):
         django.setup()
-    apps = sys.argv[1:] or ['sendgrid', ]
+    apps = sys.argv[1:] or ['django_sendgrid', ]
     test_runner = get_runner(settings)
     test_runner = test_runner(verbosity=1, interactive=True, failfast=False)
     failures = test_runner.run_tests(apps)
