+++ tox.ini
@@ -14,4 +14,4 @@ deps =
 [testenv:flake8]
 basepython=python
 deps=flake8
-commands=flake8 --max-line-length=120 sendgrid
+commands=flake8 --max-line-length=120 .
